pyproject.toml
[tool.poetry]
name = "prompti"
version = "0.1.0"
description = ""
authors = ["Manuel Gil <manuelgilsitio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
ruamel-yaml = "^0.18.6"
textstat = "^0.7.4"

[tool.poetry.scripts]
test = 'prompti.run:main'
test2 = 'prompti.run:main2'
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
